+++ lib/home/view/home_page.dart
@@ -1,14 +1,7 @@
-// ignore_for_file: use_build_context_synchronously, unawaited_futures
-
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:style_guide_infra/style_guide_infra.dart';
-import 'package:style_guide_ui/style_guide_ui.dart';
 
-import '../../l10n/l10n.dart';
 import '../../routes.dart';
-import '../../shared/cubit/selected_style_guide/selected_style_guide_cubit.dart';
-import '../../utils/bottom_sheets.dart';
 import '../cubit/all_style_guide_cubit.dart';
 
 class HomePage extends StatefulWidget {
@@ -36,7 +29,7 @@ class _HomePageState extends State<HomePage> {
             itemCount: state.styleGuides.length,
             itemBuilder: (BuildContext context, int index) {
               return ListTile(
-                title: Text(state.styleGuides[index].id.toString()),
+                title: Text(state.styleGuides[index].toString()),
               );
             },
           );
