+++ lib/shared/cubit/selected_style_guide/selected_style_guide_cubit.dart
@@ -1,8 +1,11 @@
 import 'package:bloc/bloc.dart';
 import 'package:flutter/material.dart';
 import 'package:style_guide_infra/style_guide_infra.dart';
+import 'package:style_guide_repository/style_guide_repository.dart';
 import 'package:uuid/uuid.dart';
 
+import '../../../utils/utils.dart';
+import '../../model/selected_font.dart';
 import '../../model/style_guide_model.dart';
 
 part 'selected_style_guide_state.dart';
@@ -13,29 +16,18 @@ class SelectedStyleGuideCubit extends Cubit<SelectedStyleGuideState> {
   final uuid = const Uuid();
 
   void createNewStyleGuide() {
-    emit(state.copyWith(selectedStyleGuideModel: StyleGuideModel(id: uuid.v1())));
+    print(state.toString());
   }
 
+  /// generate a default color pallet with primary color
   void createColorPallet(Color? color) {
-    final newColor = SColor(
-      primary: [color ?? AppColor.primary],
-      semantic: [AppColor.success, AppColor.error, AppColor.warn],
-      neutral: [
-        AppColor.black,
-        AppColor.grey,
-        AppColor.lightGrey1,
-        AppColor.lightGrey2,
-        AppColor.lightGrey3,
-        AppColor.lightGrey4,
-        AppColor.white,
-      ],
-      secondary: [],
-    );
-    emit(state.copyWith(selectedStyleGuideModel: state.selectedStyleGuideModel?.copyWith(color: newColor)));
+    final newColor = kDefaultColor(color);
+    emit(state.copyWith(selectedColors: newColor));
   }
 
+  /// create a new color
   void addColor({required ColorType colorType, required Color color}) {
-    final tempColor = state.selectedStyleGuideModel!.color!;
+    final tempColor = state.selectedColors!;
 
     switch (colorType) {
       case ColorType.primary:
@@ -51,11 +43,12 @@ class SelectedStyleGuideCubit extends Cubit<SelectedStyleGuideState> {
         tempColor.neutral = [...tempColor.neutral!, color];
         break;
     }
-    emit(state.copyWith(selectedStyleGuideModel: state.selectedStyleGuideModel!.copyWith(color: tempColor)));
+    emit(state.copyWith(selectedColors: tempColor));
   }
 
+  /// edit a color value
   void editColor({required ColorType colorType, required int index, required Color color}) {
-    final tempColor = state.selectedStyleGuideModel!.color!;
+    final tempColor = state.selectedColors!;
 
     switch (colorType) {
       case ColorType.primary:
@@ -71,11 +64,12 @@ class SelectedStyleGuideCubit extends Cubit<SelectedStyleGuideState> {
         tempColor.neutral?[index] = color;
         break;
     }
-    emit(state.copyWith(selectedStyleGuideModel: state.selectedStyleGuideModel!.copyWith(color: tempColor)));
+    emit(state.copyWith(selectedColors: tempColor));
   }
 
+  /// delete an color form the list based on [ColorType]
   void deleteColor({required ColorType colorType, required int index}) {
-    final tempColor = state.selectedStyleGuideModel!.color!;
+    final tempColor = state.selectedColors!;
 
     switch (colorType) {
       case ColorType.primary:
@@ -91,6 +85,63 @@ class SelectedStyleGuideCubit extends Cubit<SelectedStyleGuideState> {
         tempColor.neutral!.removeAt(index);
         break;
     }
-    emit(state.copyWith(selectedStyleGuideModel: state.selectedStyleGuideModel!.copyWith(color: tempColor)));
+    emit(state.copyWith(selectedColors: tempColor));
+  }
+
+  /// emit state when user select a font
+  /// font can be primary or secondary
+  void selectFont({required GoogleFontModel selectedFont, required bool isPrimary}) {
+    final font = SelectedFontModel(
+      fontStyle: selectedFont.family!,
+      variants: selectedFont.variants!,
+      files: selectedFont.files!,
+    );
+
+    if (isPrimary) {
+      emit(state.copyWith(primaryFont: font));
+    } else {
+      emit(state.copyWith(secondaryFont: font));
+    }
+  }
+
+  /// update selected font's weight
+  void selectFontWeight({required bool isPrimaryFont, required String fontWeight}) {
+    var tempFontWeights = <String>[];
+
+    if (isPrimaryFont) {
+      tempFontWeights = state.primaryFont!.selectedVariants;
+    } else {
+      tempFontWeights = state.secondaryFont!.selectedVariants;
+    }
+
+    // add or remove fonts
+    if (tempFontWeights.contains(fontWeight)) {
+      tempFontWeights.remove(fontWeight);
+    } else {
+      tempFontWeights = [fontWeight, ...tempFontWeights];
+    }
+
+    // emit state
+    if (isPrimaryFont) {
+      emit(
+        state.copyWith(
+          primaryFont: state.primaryFont!.copyWith(selectedVariants: tempFontWeights),
+        ),
+      );
+    } else {
+      emit(
+        state.copyWith(
+          secondaryFont: state.secondaryFont!.copyWith(selectedVariants: tempFontWeights),
+        ),
+      );
+    }
+  }
+
+  void generateDefaultFont({bool isPrimaryFont = true}) {
+    if (isPrimaryFont) {
+      emit(state.copyWith(primaryFont: kDefaultPrimaryFont));
+    } else {
+      emit(state.copyWith(secondaryFont: kDefaultSecondaryFont));
+    }
   }
 }
